refactor(twitter): decouple post class
